Decompose Router Functionality
This modification decomposes some functions in ShopifyAPI.Router so that
the underlying logic is available outside of the router context. It
also breaks down the flow of logic so that it's more self-documenting
regarding what it's doing.

Logic is moved to the ShopInstaller module and the Router module is left
to do routing type things.

The update also adds verification of the shopify store name (as
documented in Shopify documentation as an expected validation that
occurs).

Addition of shop name checking requires us to override some behavior by
environment. Because we're using a Bypass server instead of a mock,
we need to override the behavior in logic so that it handles both the
request to the alternate service while still allowing the check for the
shop name.