Integrate ShopServer Actions Into Shop Installation Flow
This modification ensures uniform behavior between the AuthTokenServer
and ShopServer upon initial installation of an App. I'm not familiar
enough with the history of these two features to know why they are
treated differently... I'm assuming that the Shop.post_install callback
is intended to handle this kind of modification; however, that relies
significantly on implementation outside of the library framework.

This modification integrates an analog to the AuthTokenServer.set() call
that's already a part of the App installation flow, ensuring that
there's also a ShopServer.set() call made on the default shop domain.

As this is a functional change to the existing flow, where we're
deciding that we also -- in every case -- want to add a Shop for every
app install, and additionally, because  I'm not really certain about the
overall desire and intent here (i.e. why Shop entries are treated
fundamentally different from AuthToken entries), I'm making this
modification a separate. I hope that I can either understand why we're
treating these differently or that we can treat these entities in a
uniform fashion during the installation process.