Allow For Args That Use The Same Pattern As create Func in Other Modules
When calling the ShopifyAPI.REST.RecurringApplicationCharge.create
function, there's an expectation that a map with the shape
%{recurring_application_charge: map()} is passed. This is not the same
as other create calls where the outer map is provided by the requesting
logic.

This update preserves the ability to call the logic using the old
pattern as well as adds a new function that provides the ability to
create RecurringApplicationCharge records in Shopify using the same
pattern that other ShopifyAPI create functions use.