ensure we handle case where user installs but doesn't accept charges
this update modifies the install flow to allow the app to determine
whether the app has been installed as a checkpoint in the flow.

This primarily allows us to look for details regarding the intallation
process that may not be available within the ShopifyAPI scope. For
example, if the shop has installed the app, but hasn't accepted the
associated recurring charge.